Added music disks
I overthought the music disc mechanic and came to the conclusion, that there should not be music disks for every song, but rather category music discs(rare, epic, legendary, etc.) and that we should be able to exchange a certain amount of rare or epic disks for a rare or epic song(10 rare discs for a rare song and 3 epic disks for an epic song). Also the EQ bars on the car stopped working. We also need to nerf the spawn rate of the disks